Make sure btrbk service only runs after network is available
As specified at
https://www.freedesktop.org/software/systemd/man/systemd.timer.html#Persistent=
the service will be activated immediately on machine boot if it would
have been triggered when the machine was turned off. This makes it
possible for the service to be activated before remote hosts are
available. Adding the `After` directive should make sure that it is
only triggered after remote hosts are reachable.