btrbk.conf.example: revert to sane non-@subvol notation
We don't want to promote weird naming conventions here.

The idea behind the change in the previous commit [1] was to get rid
of confused people asking me: "hey, I don't have a subvolume for my
rootfs, all I see is @ and @home".

Left an example for "subvolume @" which should be clear to Ubuntu
people out there.

  [1] de6c7ab586 btrbk.conf.example: use @subvol notation