btrbk: allow compressors to not have a compression level
This commit assumes that a negative `level_min`-value, which would never really
make sense as it would result in an option `--#`, means that the compression
tool does not support any level options at all.

A negative `level_max`-value would in principle still allow to have no upper
limmit, and is therefore not checked for.

Signed-off-by: Christoph Anton Mitterer <mail@christoph.anton.mitterer.name>