ssh_filter_btrbk.sh: use more common exit statuses
POSIX designates a number of exit statuses around `127` for special use and GNU
a few further.
Further, using values >127 is rather uncommon for normal use-cases.

Use `1` when the SSH command was rejected and `2` when the program’s arguments
could not be parsed).

Since this might at least in principle be used by 3rd-party programs, added a
specific note to the changelog.

Signed-off-by: Christoph Anton Mitterer <mail@christoph.anton.mitterer.name>